refactor: UI 컴포넌트 뷰와 로직 분리
