+++ src/@types/Modal.ts
@@ -1,0 +1,8 @@
+interface ModalProps {
+  isModalOpen: boolean;
+  setIsModalOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  currentPage: number;
+  setCurrentPage: React.Dispatch<React.SetStateAction<number>>;
+}
+
+export type { ModalProps };
