+++ src/components/features/AccountManageModal/DefaultInfoModal.tsx
@@ -1,98 +1,14 @@
-import { useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 import { Button } from '@chakra-ui/react';
-import toast from 'react-hot-toast';
-import {
-  useChangeBanStateRequest,
-  useGetAdminAccountRequest,
-  useGiveManagerAuthRequest,
-  useGetUserAccountRequest,
-} from '../../../api-hooks/Account';
 import { DefaultButton } from '../../ui/ButtonUI';
 import { AccountInfoText, FooterSection } from './style';
-import { IAccountData } from '../../../@types/Account';
-
-export interface SelectedAccountProps {
-  selectedAccount: IAccountData;
-  setIsModalOpen: React.Dispatch<React.SetStateAction<boolean>>;
-  currentPage: number;
-  setCurrentPage: React.Dispatch<React.SetStateAction<number>>;
-}
+import { SelectedAccountProps } from '../../../@types/Account';
+import useDefaultInfoModal from './useDefaultInfoModal';
 
 function DefaultInfo({ selectedAccount, setIsModalOpen, currentPage, setCurrentPage }: SelectedAccountProps) {
   const location = useLocation();
-
-  const { adminAccountRefetch } = useGetAdminAccountRequest({ page: currentPage - 1 });
-  const { accountRefetch } = useGetUserAccountRequest({ page: currentPage - 1 });
-
-  const {
-    mutate: accountBanStateRequest,
-    data: accountBanStateData,
-    error: accountBanStateError,
-    isLoading: accountBanStateLoading,
-  } = useChangeBanStateRequest();
-
-  const {
-    mutate: giveManagerAuthRequest,
-    data: giveManagerAuthStateData,
-    error: giveManagerAuthStateError,
-    isLoading: giveManagerAuthStateLoading,
-  } = useGiveManagerAuthRequest();
-
-  useEffect(() => {
-    if (accountBanStateData) {
-      if (selectedAccount.ban) {
-        toast.success('계정이 활성화되었습니다.');
-        setCurrentPage(1);
-      } else if (!selectedAccount.ban) {
-        toast('계정이 비활성화되었습니다.', {
-          icon: '⚠️',
-        });
-      }
-      void adminAccountRefetch();
-      void accountRefetch();
-      setIsModalOpen(false);
-    } else if (accountBanStateError) {
-      toast.error((accountBanStateError as Error).message);
-    }
-  }, [accountBanStateData, accountBanStateError]);
-
-  useEffect(() => {
-    if (giveManagerAuthStateData) {
-      if (!selectedAccount.manager) {
-        toast.success('해당 계정에 매니저 권한이 부여되었습니다.');
-        setCurrentPage(1);
-      } else if (selectedAccount.manager) {
-        toast('해당 계정의 매니저 권한이 제거되었습니다.', {
-          icon: '⚠️',
-        });
-      }
-      void adminAccountRefetch();
-      setIsModalOpen(false);
-    } else if (giveManagerAuthStateError) {
-      toast.error((giveManagerAuthStateError as Error).message);
-    }
-  }, [giveManagerAuthStateData, giveManagerAuthStateError]);
-
-  const handleChangeBanState = () => {
-    if (selectedAccount.ban) {
-      const isYesClicked = confirm('정말 해당 계정을 활성화 하시겠습니까?');
-      if (isYesClicked) accountBanStateRequest({ id: selectedAccount.id });
-    } else if (!selectedAccount.ban) {
-      const isYesClicked = confirm('정말 해당 계정을 비활성화 하시겠습니까?');
-      if (isYesClicked) accountBanStateRequest({ id: selectedAccount.id });
-    }
-  };
-
-  const handleGiveManagerAuth = () => {
-    if (!selectedAccount.manager) {
-      const isYesClicked = confirm('정말 해당 계정에 매니저 권한을 부여하시겠습니까?');
-      if (isYesClicked) giveManagerAuthRequest({ id: selectedAccount.id });
-    } else if (selectedAccount.manager) {
-      const isYesClicked = confirm('정말 해당 계정의 매니저 권한을 해제하시겠습니까??');
-      if (isYesClicked) giveManagerAuthRequest({ id: selectedAccount.id });
-    }
-  };
+  const { accountBanStateLoading, giveManagerAuthStateLoading, handleChangeBanState, handleGiveManagerAuth } =
+    useDefaultInfoModal({ selectedAccount, setIsModalOpen, currentPage, setCurrentPage });
 
   return (
     <div>
