+++ src/components/features/AlertDialogModal/index.tsx
@@ -1,4 +1,4 @@
-import { useRef, useEffect } from 'react';
+import { useRef } from 'react';
 import {
   AlertDialog,
   AlertDialogBody,
@@ -9,17 +9,9 @@ import {
   AlertDialogCloseButton,
   Button,
 } from '@chakra-ui/react';
-import { useDeleteCategoryRequest, useGetCategoryRequest } from '../../../api-hooks/InfoCategory';
-import { useDeleteInfoRequest, useGetInfoRequest } from '../../../api-hooks/Info';
-import { useDeleteNoticeRequest, useGetNoticeRequest } from '../../../api-hooks/Notice';
-import { useDeleteBuildingRequest, useGetBuildingRequest } from '../../../api-hooks/Building';
-import { useGetMyArticleRequest, useGetMyNoticeRequest } from '../../../api-hooks/MyPage';
-import toast from 'react-hot-toast';
-import { useSelectedArticleAtom } from '../../../store/articleAtom';
-import { useSelectedBuildingAtom } from '../../../store/buildingAtom';
-import { useCurrentLocationAtom } from '../../../store/currentLocationAtom';
+import useAlertDialogModal from './useAlertDialogModal';
 
-interface AlertProps {
+export interface AlertProps {
   location?: string;
   selectedItemIndex?: number;
   isAlertOpen: boolean;
@@ -37,139 +29,14 @@ function AlertDialogModal({
   setCurrentPage,
 }: AlertProps) {
   const cancelRef = useRef(null);
-  const { selectedArticle, resetAtom } = useSelectedArticleAtom();
-  const { selectedBuilding, resetBuildingAtom } = useSelectedBuildingAtom();
-  const { currentLocation } = useCurrentLocationAtom();
-
-  const { getCategoryRefetch } = useGetCategoryRequest(
-    {
-      page: currentPage - 1,
-    },
-    false,
-  );
-
-  const { infoRefetch } = useGetInfoRequest(
-    {
-      page: currentPage - 1,
-      startDate: '',
-      endDate: '',
-      title: '',
-      typeName: '',
-    },
-    false,
-  );
-
-  const { noticeRefetch } = useGetNoticeRequest(
-    {
-      page: currentPage - 1,
-      startDate: '',
-      endDate: '',
-      title: '',
-    },
-    false,
-  );
-
-  const { getBuildingRefetch } = useGetBuildingRequest(
-    {
-      page: currentPage - 1,
-    },
-    false,
-  );
-
-  const { myArticleRefetch } = useGetMyArticleRequest(
-    {
-      page: currentPage - 1,
-    },
-    false,
-  );
-
-  const { myNoticeRefetch } = useGetMyNoticeRequest(
-    {
-      page: currentPage - 1,
-    },
-    false,
-  );
-
-  const {
-    mutate: categoryDeleteRequest,
-    data: categoryDeleteData,
-    error: categoryDeleteError,
-  } = useDeleteCategoryRequest();
-
-  const { mutate: infoDeleteRequest, data: infoDeleteData, error: infoDeleteError } = useDeleteInfoRequest();
-
-  const { mutate: noticeDeleteRequest, data: noticeDeleteData, error: noticeDeleteError } = useDeleteNoticeRequest();
-
-  const {
-    mutate: deleteBuildingRequest,
-    data: deleteBuildingData,
-    error: deleteBuildingError,
-  } = useDeleteBuildingRequest();
-
-  const handleAlertDialog = () => {
-    setIsAlertOpen(false);
-  };
-
-  const handleDelete = () => {
-    if (location === '카테고리') {
-      categoryDeleteRequest({ id: selectedItemIndex });
-    } else if (location === '정보' || currentLocation === '작성한 정보') {
-      infoDeleteRequest({ id: selectedArticle.id });
-    } else if (location === '공지' || currentLocation === '작성한 공지사항') {
-      noticeDeleteRequest({ id: selectedArticle.id });
-    } else if (location === '건물') {
-      deleteBuildingRequest({ id: selectedBuilding.id });
-    }
-  };
-
-  useEffect(() => {
-    if (categoryDeleteData) {
-      toast.success('삭제되었습니다.');
-      void getCategoryRefetch();
-      setCurrentPage(1);
-    } else if (categoryDeleteError) {
-      toast.error((categoryDeleteError as Error).message);
-    }
-    setIsAlertOpen(false);
-  }, [categoryDeleteData, categoryDeleteError]);
-
-  useEffect(() => {
-    if (infoDeleteData) {
-      toast.success('삭제되었습니다.');
-      void infoRefetch();
-      void myArticleRefetch();
-      setCurrentPage(1);
-      resetAtom();
-    } else if (infoDeleteError) {
-      toast.error((infoDeleteError as Error).message);
-    }
-    setIsAlertOpen(false);
-  }, [infoDeleteData, infoDeleteError]);
-
-  useEffect(() => {
-    if (noticeDeleteData) {
-      toast.success('삭제되었습니다.');
-      void noticeRefetch();
-      void myNoticeRefetch();
-      setCurrentPage(1);
-      resetAtom();
-    } else if (noticeDeleteError) {
-      toast.error((noticeDeleteError as Error).message);
-    }
-    setIsAlertOpen(false);
-  }, [noticeDeleteData, noticeDeleteError]);
-
-  useEffect(() => {
-    if (deleteBuildingData) {
-      toast.success('삭제되었습니다.');
-      void getBuildingRefetch();
-      setCurrentPage(1);
-      resetBuildingAtom();
-    } else if (deleteBuildingError) {
-      toast.error((deleteBuildingError as Error).message);
-    }
-    setIsAlertOpen(false);
-  }, [deleteBuildingData, deleteBuildingError]);
+  const { handleAlertDialog, handleDelete } = useAlertDialogModal({
+    location,
+    selectedItemIndex,
+    isAlertOpen,
+    setIsAlertOpen,
+    currentPage,
+    setCurrentPage,
+  });
 
   return (
     <AlertDialog
@@ -202,19 +69,15 @@ export default AlertDialogModal;
 export const BanAlertDialogModal = ({ isAlertOpen, setIsAlertOpen }: AlertProps) => {
   const cancelRef = useRef(null);
 
-  const handleBanAlertDialog = () => {
-    setIsAlertOpen(false);
-  };
-
   const handleAccountBan = () => {
-    // 계정 정지하기
+    // 계정 정지하기 요청
   };
 
   return (
     <AlertDialog
       motionPreset="slideInBottom"
       leastDestructiveRef={cancelRef}
-      onClose={() => handleBanAlertDialog()}
+      onClose={() => setIsAlertOpen(false)}
       isOpen={isAlertOpen}
       isCentered
     >
@@ -223,7 +86,7 @@ export const BanAlertDialogModal = ({ isAlertOpen, setIsAlertOpen }: AlertProps)
         <AlertDialogHeader>해당 계정을 정지하시겠습니까?</AlertDialogHeader>
         <AlertDialogCloseButton />
         <AlertDialogFooter>
-          <Button ref={cancelRef} onClick={() => handleBanAlertDialog()}>
+          <Button ref={cancelRef} onClick={() => setIsAlertOpen(false)}>
             취소
           </Button>
           <Button colorScheme="red" ml={3} onClick={() => handleAccountBan()}>
